+++ packages/client/generate-timer-worker.sh
@@ -1,0 +1,16 @@
+#!/usr/bin/env bash
+
+npx tsc src/timers/worker.ts \
+  --skipLibCheck \
+  --removeComments \
+  --module preserve \
+  --lib ES2020,WebWorker \
+  --outDir worker-dist
+
+cat <<EOF >src/timers/worker.build.ts
+export const timerWorker = {
+  src: \`$(<worker-dist/worker.js)\`,
+};
+EOF
+
+rm -r worker-dist